+++ lib/inline_forms/form_elements/slider_with_values_helper.rb
@@ -5,7 +5,11 @@ module InlineForms
       module_eval(<<~'INLINE_FORMS_FORM_ELEMENT', __FILE__, __LINE__ + 1)
     # -*- encoding : utf-8 -*-
     
-    # slider_with_values
+    # slider_with_values -- native <input type="range"> since 8.1.26 (replaces
+    # the jQuery UI slider and its inline <script>). The label text next to the
+    # slider is kept in sync by initInlineFormsWidgets (class hook
+    # `slider_with_values`, labels in data-slider-values, target <output> id in
+    # data-slider-output); helpers emit no per-field scripts.
     def slider_with_values_show(object, attribute)
       values = attribute_values(object, attribute)
       value = object.send(attribute).to_i         # should be an int
@@ -14,26 +18,17 @@ module InlineForms
       if value == 0
         out = "?"   # we use this as the 'unknown' value. So in the data, 0 should always be the unknown value. This gives problems with sliders where the real value is 0.
       else
+        # <progress> (not a disabled <input type=range>): the show panel is
+        # wrapped in the inline-edit link, and a disabled input would swallow
+        # the click; progress is non-interactive so the click-to-edit works.
         out = "".html_safe
-        out << "<div class='slider slider_#{attribute.to_s}' id='slider_#{css_id}'></div>".html_safe
-        out << "<div class='slider_value' id='value_#{css_id}'>#{display_value}</div>".html_safe
-        out << "<div style='clear: both' />".html_safe
-        out << "<input type='hidden' name='_#{object.class.to_s.underscore}[#{attribute}]' value='0' id='input_#{css_id}' />".html_safe
-        out << ('<script>
-                  $(function() {
-                    var displayvalues = ' + values.collect {|x| x[1]}.inspect + ';
-                    $( "#slider_' + css_id + '" ).slider(
-                      {
-                        value:' + value.to_s + ',
-                        range: "min",
-                        disabled: true,
-                        min: 0,
-                        max: ' + values.length.to_s + ',
-                        step: 1,
-                      }
-                    );
-                  });
-                 </script>').html_safe
+        out << content_tag(:progress, nil,
+                           value: value,
+                           max: values.collect(&:first).max.to_i,
+                           class: "slider slider_#{attribute}",
+                           id: "slider_#{css_id}")
+        out << content_tag(:div, display_value, class: 'slider_value', id: "value_#{css_id}")
+        out << "<div style='clear: both'></div>".html_safe
       end
       link_to_inline_edit object, attribute, out, from_callee: __callee__
     end
@@ -44,55 +39,23 @@ module InlineForms
       value = object.send(attribute).to_i         # should be an int, will be 0 if nil
       css_id = "#{object.class.to_s.underscore}_#{object.id}_#{attribute}"
       display_value = values.assoc(value)[1]      # values should be [ [ 0, value ], [ 3, value2 ] .... ] and we lookup the key, not the place in the array!
+      labels = values.collect { |x| x[1] }
       out = "".html_safe
-      out << "<div class='slider slider_#{attribute.to_s}' id='slider_#{css_id}'></div>".html_safe
-      out << "<div class='slider_value' id='value_#{css_id}'>#{display_value}</div>".html_safe
-      out << "<div style='clear: both' />".html_safe
-      out << "<input type='hidden' name='_#{object.class.to_s.underscore}[#{attribute}]' value='0' id='input_#{css_id}' />".html_safe
-      out << ('<script>
-    	$(function() {
-        var displayvalues = ' + values.collect {|x| x[1]}.inspect + ';
-    		$( "#slider_' + css_id + '" ).slider(
-          {
-            value:' + value.to_s + ',
-            range: "min",
-            min: 0,
-            max: 5,
-            step: 1,
-            slide: function( event, ui ) {
-              $( "#input_' + css_id + '" ).val( ui.value );
-              $( "#value_' + css_id + '" ).html( displayvalues[ui.value] );
-            }
-      		}
-        );').html_safe
-      out << ('$( "#value_' + css_id + '" ).html(displayvalues[' + value.to_s + ']);').html_safe
-    	out << ('$( "#input_' + css_id + '" ).val(' + value.to_s + ');').html_safe
-    	out << ('});
-    	</script>').html_safe
+      out << range_field_tag("_#{object.class.to_s.underscore}[#{attribute}]", value,
+                             id: "input_#{css_id}",
+                             class: "slider slider_#{attribute} slider_with_values",
+                             in: 0..(values.collect(&:first).max.to_i),
+                             step: 1,
+                             data: { slider_values: labels.to_json,
+                                     slider_output: "value_#{css_id}" })
+      out << content_tag(:output, display_value, class: 'slider_value', id: "value_#{css_id}")
+      out << "<div style='clear: both'></div>".html_safe
       out
     end
     
     def slider_with_values_update(object, attribute)
       object[attribute.to_sym] = params[('_' + object.class.to_s.underscore).to_sym][attribute.to_sym]
     end
-    
-    #<style>
-    #	#demo-frame > div.demo { padding: 10px !important; };
-    #	</style>
-    #
-    #
-    #
-    #<div class="demo">
-    #
-    #<p>
-    #	<label for="amount">Donation amount ($50 increments):</label>
-    #	<input type="text" id="amount" style="border:0; color:#f6931f; font-weight:bold;" />
-    #</p>
-    #
-    #<div id="slider"></div>
-    #
-    #</div><!-- End demo -->
-    #
       INLINE_FORMS_FORM_ELEMENT
     end
   end